Rename scenes file to sceneManager
