+++ src/game/extras.ts
@@ -3,7 +3,7 @@ import * as THREE from 'three';
 import { getRandom, getHtmlColor } from '../utils/lba';
 import { SpriteType } from './data/spriteType';
 import { loadSprite } from '../iso/sprites';
-import { addExtraToScene, removeExtraFromScene } from './scenes';
+import { addExtraToScene, removeExtraFromScene } from './sceneManager';
 import { clone } from 'lodash';
 import { MAX_LIFE } from './gameState';
 // import { createBoundingBox } from '../utils/rendering';
