+++ src/game/sceneManager.ts
