+++ src/ui/game/TeleportMenu.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {map, filter} from 'lodash';
 import LocationsNode from '../editor/areas/gameplay/locator/LocationsNode';
-import { SceneManager } from '../../game/scenes';
+import { SceneManager } from '../../game/sceneManager';
 import { Game } from '../../game/game';
 
 const style = {
